+++ src/function/chat_message.ts
@@ -1,3 +1,5 @@
+import { highlight_code } from '@/util/highlight_code';
+
 import {
   addOneMessage,
   chat,
@@ -179,6 +181,22 @@ export async function setChatMessages(
   chat_messages: Array<{ message_id: number } & (Partial<ChatMessage> | Partial<ChatMessageSwiped>)>,
   { refresh = 'affected' }: SetChatMessagesOption = {},
 ): Promise<void> {
+  const convert_and_merge_messages = (
+    data: Array<{ message_id: number } & (Partial<ChatMessage> | Partial<ChatMessageSwiped>)>,
+  ): any => {
+    return _(data)
+      .map(chat_message => ({
+        ...chat_message,
+        message_id: chat_message.message_id < 0 ? chat.length + chat_message.message_id : chat_message.message_id,
+      }))
+      .sortBy('message_id')
+      .groupBy('message_id')
+      .map(messages => {
+        return messages.reduce((result, current) => ({ ...result, ...current }), {});
+      })
+      .value();
+  };
+
   const is_chat_message = (
     chat_message: { message_id: number } & (Partial<ChatMessage> | Partial<ChatMessageSwiped>),
   ): chat_message is { message_id: number } & Partial<ChatMessage> => {
@@ -282,12 +300,16 @@ export async function setChatMessages(
           message_id,
         ),
       );
+    $mes_html.find('pre code').each((_index, element) => {
+      highlight_code(element);
+    });
     await eventSource.emit(
       chat_message.is_user ? event_types.USER_MESSAGE_RENDERED : event_types.CHARACTER_MESSAGE_RENDERED,
       message_id,
     );
   };
 
+  chat_messages = convert_and_merge_messages(chat_messages);
   await Promise.all(chat_messages.map(modify));
   await saveChatConditional();
   if (refresh === 'all') {